fix: Fix Like count logic

See merge request s09-webmobile1-sub2/S09P12A109!260

+++ Backend/src/main/java/com/ssafy/mozzi/api/response/MozzirollLikeRes.java
@@ -13,6 +13,6 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 public class MozzirollLikeRes {
-    private int likeCount;
+    private long likeCount;
     private boolean isLiked;
 }
